agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.12.5",
    "pydantic-ai>=1.39.0",
    "python-dotenv>=1.2.1",
]
